Almost done with loading into custom levels
Stuck working on NPlayerStateController.ActivateLevelPlayer(), line 35 "NullReferenceException"

+++ NiGHTS Open Dreams/Assets/Scripts/Player/Updated Scripts/NPlayerScriptableObject.cs
@@ -89,5 +89,5 @@ public class NPlayerScriptableObject : ScriptableObject
 
     public float displayScore;
     public float levelTime;
-    public float openChips;    
+    public int openChips;    
 }
